# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for FAQ accordion.",
        "title": "FAQ Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/faqs": {
            "get": {
                "description": "Get active FAQs ordered by position",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faqs"
                ],
                "summary": "List FAQs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.FAQListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new FAQ item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faqs"
                ],
                "summary": "Create FAQ",
                "parameters": [
                    {
                        "description": "FAQ payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateFAQRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.FAQResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/faqs/{id}": {
            "get": {
                "description": "Get one FAQ by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faqs"
                ],
                "summary": "Get FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.FAQResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update FAQ by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faqs"
                ],
                "summary": "Update FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "FAQ payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateFAQRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.FAQResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete FAQ by id (idempotent)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faqs"
                ],
                "summary": "Delete FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateFAQRequest": {
            "description": "CreateFAQRequest describes request body for creating a FAQ.",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.ErrorResponse": {
            "description": "ErrorResponse describes an error payload.",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "domain.FAQFullResponse": {
            "description": "FAQFullResponse is a full FAQ representation.",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.FAQListItemResponse": {
            "description": "FAQListItemResponse is a short FAQ representation used in lists.",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.FAQListResponse": {
            "description": "FAQListResponse wraps a list response.",
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.FAQListItemResponse"
                    }
                }
            }
        },
        "domain.FAQResponse": {
            "description": "FAQResponse wraps a single FAQ response.",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.FAQFullResponse"
                }
            }
        },
        "domain.MessageResponse": {
            "description": "MessageResponse is a simple message response.",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.UpdateFAQRequest": {
            "description": "UpdateFAQRequest describes request body for updating a FAQ.",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}